+++ clang/lib/Sema/SemaDecl.cpp
@@ -14333,6 +14333,14 @@ void Sema::ActOnUninitializedDecl(Decl *RealDecl) {
       return;
     }
 
+    // In C, if the definition is const-qualified and has no initializer, it
+    // is left uninitialized unless it has static or thread storage duration.
+    if (!getLangOpts().CPlusPlus && Type.isConstQualified())
+      Diag(Var->getLocation(), diag::warn_default_init_const)
+          << Type << /*not a field*/0
+          << (Var->getStorageDuration() != SD_Static &&
+              Var->getStorageDuration() != SD_Thread);
+
     // Check for jumps past the implicit initializer.  C++0x
     // clarifies that this applies to a "variable with automatic
     // storage duration", not a "local variable".